backend/requirements.txt
# backend/requirements.txt

# Core da Aplicação
django==5.2.3
gunicorn==22.0.0

# Banco de Dados
mysqlclient==2.2.7

# API e Comunicação
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Autenticação por Token
djangorestframework-simplejwt==5.3.1

# Variáveis de Ambiente
python-dotenv==1.0.1